+++ kubernetes-model-generator/kubernetes-model-coordination/src/main/java/io/fabric8/kubernetes/api/model/coordination/v1/LeaseSpec.java
@@ -199,6 +199,7 @@ public class LeaseSpec implements Editable<LeaseSpecBuilder>, KubernetesResource
   }
 
   @JsonAnyGetter
+  @JsonIgnore
   public Map<String, Object> getAdditionalProperties() {
     return this.additionalProperties;
   }